Update systed-boot update and no format on manual partition

+++ alis.sh
@@ -472,44 +472,46 @@ function partition() {
     fi
 
     # format
-    # Delete patition filesystem in case is reinstalling in an already existing system
-    # Not fail on error
-    wipefs -a -f "$PARTITION_BOOT" || true
-    wipefs -a -f "$DEVICE_ROOT" || true
+    if [ "$PARTITION_MODE" != "manual" ]; then
+        # Delete patition filesystem in case is reinstalling in an already existing system
+        # Not fail on error
+        wipefs -a -f "$PARTITION_BOOT" || true
+        wipefs -a -f "$DEVICE_ROOT" || true
 
-    ## boot
-    if [ "$BIOS_TYPE" == "uefi" ]; then
-        mkfs.fat -n ESP -F32 "$PARTITION_BOOT"
-    fi
-    if [ "$BIOS_TYPE" == "bios" ]; then
-        mkfs.ext4 -L boot "$PARTITION_BOOT"
-    fi
-    ## root
-    if [ "$FILE_SYSTEM_TYPE" == "reiserfs" ]; then
-        mkfs."$FILE_SYSTEM_TYPE" -f -l root "$DEVICE_ROOT"
-    elif [ "$FILE_SYSTEM_TYPE" == "f2fs" ]; then
-        mkfs."$FILE_SYSTEM_TYPE" -l root "$DEVICE_ROOT"
-    else
-        mkfs."$FILE_SYSTEM_TYPE" -L root "$DEVICE_ROOT"
-    fi
-    ## mountpoint
-    for I in "${PARTITION_MOUNT_POINTS[@]}"; do
-        if [[ "$I" =~ ^!.* ]]; then
-            continue
+        ## boot
+        if [ "$BIOS_TYPE" == "uefi" ]; then
+            mkfs.fat -n ESP -F32 "$PARTITION_BOOT"
         fi
-        IFS='=' read -ra PARTITION_MOUNT_POINT <<< "$I"
-        if [ "${PARTITION_MOUNT_POINT[1]}" == "/boot" ] || [ "${PARTITION_MOUNT_POINT[1]}" == "/" ]; then
-            continue
+        if [ "$BIOS_TYPE" == "bios" ]; then
+            mkfs.ext4 -L boot "$PARTITION_BOOT"
         fi
-        local PARTITION_DEVICE="$(partition_device "$DEVICE" "${PARTITION_MOUNT_POINT[0]}")"
+        ## root
         if [ "$FILE_SYSTEM_TYPE" == "reiserfs" ]; then
-            mkfs."$FILE_SYSTEM_TYPE" -f "$PARTITION_DEVICE"
+            mkfs."$FILE_SYSTEM_TYPE" -f -l root "$DEVICE_ROOT"
         elif [ "$FILE_SYSTEM_TYPE" == "f2fs" ]; then
-            mkfs."$FILE_SYSTEM_TYPE" "$PARTITION_DEVICE"
+            mkfs."$FILE_SYSTEM_TYPE" -l root "$DEVICE_ROOT"
         else
-            mkfs."$FILE_SYSTEM_TYPE" "$PARTITION_DEVICE"
+            mkfs."$FILE_SYSTEM_TYPE" -L root "$DEVICE_ROOT"
         fi
-    done
+        ## mountpoint
+        for I in "${PARTITION_MOUNT_POINTS[@]}"; do
+            if [[ "$I" =~ ^!.* ]]; then
+                continue
+            fi
+            IFS='=' read -ra PARTITION_MOUNT_POINT <<< "$I"
+            if [ "${PARTITION_MOUNT_POINT[1]}" == "/boot" ] || [ "${PARTITION_MOUNT_POINT[1]}" == "/" ]; then
+                continue
+            fi
+            local PARTITION_DEVICE="$(partition_device "$DEVICE" "${PARTITION_MOUNT_POINT[0]}")"
+            if [ "$FILE_SYSTEM_TYPE" == "reiserfs" ]; then
+                mkfs."$FILE_SYSTEM_TYPE" -f "$PARTITION_DEVICE"
+            elif [ "$FILE_SYSTEM_TYPE" == "f2fs" ]; then
+                mkfs."$FILE_SYSTEM_TYPE" "$PARTITION_DEVICE"
+            else
+                mkfs."$FILE_SYSTEM_TYPE" "$PARTITION_DEVICE"
+            fi
+        done
+    fi
 
     # options
     partition_options
@@ -1340,9 +1342,9 @@ Operation = Upgrade
 Target = systemd
 
 [Action]
-Description = Updating systemd-boot
+Description = Updating systemd-boot...
 When = PostTransaction
-Exec = /usr/bin/bootctl update
+Exec = /usr/bin/systemctl restart systemd-boot-update.service
 EOT
 
     local SYSTEMD_MICROCODE=""
